Use trait for dealing with svg attributes
