+++ src/Icon.php
@@ -4,9 +4,9 @@ namespace Feather;
 
 class Icon
 {
-    private $name;
+    use SvgAttributesTrait;
 
-    private $attributes;
+    private $name;
 
     private $content;
 
@@ -20,32 +20,6 @@ class Icon
     public function getName(): string
     {
         return $this->name;
-    }
-
-    public function setAttributes(array $attributes): self
-    {
-        $this->attributes = \array_merge($this->attributes, $attributes);
-
-        return $this;
-    }
-
-    public function setAttribute(string $key, $value): self
-    {
-        $this->attributes[$key] = $value;
-
-        return $this;
-    }
-
-    public function removeAttribute(string $key): self
-    {
-        unset($this->attributes[$key]);
-
-        return $this;
-    }
-
-    public function getAttributes(): array
-    {
-        return $this->attributes;
     }
 
     public function render(): string
